fix: error SSR

+++ src/shared/services/local-storage/local-storage.service.ts
@@ -2,21 +2,33 @@ import { LocalStorageService } from "./local-storage.model";
 
 export function useLocalStorage(): LocalStorageService {
   function get<T = any>(key: string): T {
-    const data = localStorage.getItem(key) || "";
-    return JSON.parse(data);
+    if (process.client) {
+      const data = localStorage.getItem(key) || "";
+      return JSON.parse(data) as T;
+    }
+
+    return {} as T;
   }
 
   function set(key: string, value: any): void {
-    const data = JSON.stringify(value);
-    localStorage.setItem(key, data);
+    if (process.client) {
+      const data = JSON.stringify(value);
+      localStorage.setItem(key, data);
+    }
   }
 
   function reset(): void {
-    localStorage.clear();
+    if (process.client) {
+      localStorage.clear();
+    }
   }
 
   function isExist(key: string): boolean {
-    return !!localStorage.getItem(key);
+    if (process.client) {
+      return !!localStorage.getItem(key);
+    }
+
+    return false;
   }
 
   return { get, isExist, reset, set };
